Move LS event to `src/gui`.
